+++ lib/components/component.dart
@@ -60,6 +60,9 @@ abstract class Component {
   /// This can be used to make initializations on your component as, when this method is called,
   /// things like resize (and other mixins) are already set and usable.
   void onMount() {}
+
+  /// Called right before the component is destroyed and removed from the game
+  void onDestroy() {}
 }
 
 /// A [Component] implementation that represents a component that has a specific, possibly dynamic position on the screen.
@@ -154,6 +157,7 @@ abstract class PositionComponent extends Component {
 /// This is the most commonly used child of [Component].
 class SpriteComponent extends PositionComponent {
   Sprite sprite;
+  Paint overridePaint;
 
   SpriteComponent();
 
@@ -171,7 +175,8 @@ class SpriteComponent extends PositionComponent {
   @override
   void render(Canvas canvas) {
     prepareCanvas(canvas);
-    sprite.render(canvas, width: width, height: height);
+    sprite.render(canvas,
+        width: width, height: height, overridePaint: overridePaint);
   }
 
   @override